% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{as_functional.matrix}
\alias{as_functional.matrix}
\title{Convert a Matrix to a functional vector}
\usage{
\method{as_functional}{matrix}(x, args = NULL, ids = NULL, ...)
}
\arguments{
\item{args}{(\code{numeric()})}

\item{ids}{(any)\cr
The ids, that assign labels to the rows of a matrix. Default is \code{1:nrow(x)}.}
}
\description{
Converts a matrix to a vector of class \code{functional}.
Rows contain various function evaluations for an individual.
}
