% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{as_functional}
\alias{as_functional}
\title{Converts object to class 'functional'}
\usage{
as_functional(x, ...)
}
\arguments{
\item{x}{(any)}
}
\description{
Convenience S3 generic that allows to construct vectors of class 'functional' from objects
types in which functional data is usually stored (e.g. \code{data.table()} or \code{matrix()}).
}
