% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dti}
\alias{dti}
\title{The dti dataset}
\format{
A data frame with 340 rows and 5 variables
}
\usage{
data(dti)
}
\description{
See \link{mlr_tasks_dti} for a description of the dataset.
}
\keyword{datasets}
\keyword{internal}
