% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten_functional}
\alias{flatten_functional}
\title{Flattens a functional vector}
\usage{
flatten_functional(x)
}
\arguments{
\item{x}{(\code{character(1)}) Prefix for the column names.}
}
\value{
Returns a matrix.
}
\description{
For each invividual ${i_1, ... i_n}$ we have obsrevations $o_j = (o_{j, 1}, ..., o_{j, n(j)})$.
Each $o_{i, j}$ is a argument-value pair $(a_{i, j}, v_{i, j})$.
Flattening a vector means creating a matrix, where there is a different column for
each distinct input $v_{i, j}$
Flattening a functional vector means creating a matrix with dimension (n, )
}
\examples{
ids = c("a", "a", "b")
args = 1:3
vals = args * 2
funct = functional(args, vals, ids)
ffunct = flatten_functional(funct)
print(ffunct)

}
