% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fuel}
\alias{fuel}
\title{The fuel dataset}
\format{
A data frame with 129 rows and 4 variables
}
\usage{
data(fuel)
}
\description{
See \link{mlr_tasks_fuel} for a description of the dataset.
}
\keyword{datasets}
\keyword{internal}
