% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpFDADepth.R
\name{mlr_pipeops_fda.depth}
\alias{mlr_pipeops_fda.depth}
\alias{PipeOpFDADepth}
\title{Functional Data Depth Features}
\description{
Computes the data depth of functional features via \code{\link[tf:tf_depth]{tf::tf_depth()}}.
Data depth measures how central each curve is relative to the others: values close to \code{1} indicate central curves
and values close to \code{0} indicate extreme curves.

The depth is computed in-sample, i.e. each curve is scored relative to the other curves in the same data. The same
operation is applied during training and prediction.

Irregular curves are interpolated to a common grid before the depth is computed. Curves that remain incomplete
after interpolation (e.g. because they only cover part of the grid) are assigned a depth of \code{NA}.
}
\details{
The \code{"MHI"} method ranks functions from lowest (\code{0}) to highest (\code{1}) instead of from most extreme to most central.
The \code{"RPD"} method relies on random projections, so set a seed (e.g. via \code{\link[=set.seed]{set.seed()}}) before training for
reproducible results.
}
\section{State}{

The \verb{$state} is a named \code{list} with the \verb{$state} elements inherited from
\code{\link[mlr3pipelines:PipeOpTaskPreproc]{PipeOpTaskPreproc}}.
}

\section{Parameters}{

The parameters are the parameters inherited from \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{PipeOpTaskPreprocSimple}},
as well as the following parameters:
\itemize{
\item \code{method} :: \code{character(1)}\cr
The depth method to use. One of \code{"MBD"} (modified band depth, the default), \code{"MHI"} (modified hypograph index),
\code{"FM"} (Fraiman-Muniz), \code{"FSD"} (functional spatial depth), or \code{"RPD"} (regularized projection depth).
Initial value is \code{"MBD"}.
\item \code{na.rm} :: \code{logical(1)}\cr
Whether to remove missing observations before computing the depth. Initial value is \code{TRUE}.
\item \code{u} :: \code{numeric(1)}\cr
Quantile level for the regularization. Only used when \code{method} is \code{"RPD"}. Default is \code{0.01}.
\item \code{n_projections} :: \code{integer(1)}\cr
Number of projection directions. Only used when \code{method} is \code{"RPD"}. Default is \code{5000}.
\item \code{n_projections_beta} :: \code{integer(1)}\cr
Number of directions for estimating the regularization parameter. Only used when \code{method} is \code{"RPD"}.
Default is \code{500}.
}
}

\section{Naming}{

The new names generally append a \verb{_depth} to the corresponding column name.
If a column was called \code{"x"}, the corresponding new column will be called \code{"x_depth"}.
}

\examples{
task = tsk("fuel")
po_depth = po("fda.depth", method = "MBD")
task_depth = po_depth$train(list(task))[[1L]]
task_depth$data(cols = c("NIR_depth", "UVVIS_depth"))
}
\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}} -> \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{mlr3pipelines::PipeOpTaskPreprocSimple}} -> \code{PipeOpFDADepth}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PipeOpFDADepth-initialize}{\code{PipeOpFDADepth$new()}}
    \item \href{#method-PipeOpFDADepth-clone}{\code{PipeOpFDADepth$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpFDADepth-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpFDADepth-initialize}{}}}
\subsection{\code{PipeOpFDADepth$new()}}{
  Initializes a new instance of this Class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PipeOpFDADepth$new(id = "fda.depth", param_vals = list())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{(\code{character(1)})\cr
Identifier of resulting object, default \code{"fda.depth"}.}
      \item{\code{param_vals}}{(named \code{list()})\cr
List of hyperparameter settings, overwriting the hyperparameter settings that would
otherwise be set during construction. Default \code{list()}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpFDADepth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpFDADepth-clone}{}}}
\subsection{\code{PipeOpFDADepth$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PipeOpFDADepth$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
