% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpFDADerive.R
\name{mlr_pipeops_fda.derive}
\alias{mlr_pipeops_fda.derive}
\alias{PipeOpFDADerive}
\title{Derivatives of Functional Columns}
\description{
Computes derivatives of functional features via \code{\link[tf:tf_derive]{tf::tf_derive()}}.
For \code{tfd} inputs derivatives are obtained by finite differencing of the function evaluations,
for \code{tfb} inputs by finite differencing of the basis functions.
}
\section{State}{

The \verb{$state} is a named \code{list} with the \verb{$state} elements inherited from
\code{\link[mlr3pipelines:PipeOpTaskPreproc]{PipeOpTaskPreproc}}.
}

\section{Parameters}{

The parameters are the parameters inherited from \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{PipeOpTaskPreprocSimple}},
as well as the following parameters:
\itemize{
\item \code{order} :: \code{integer(1)}\cr
Order of the derivative. Must be a positive integer. Initial value is \code{1}.
\item \code{arg} :: \code{numeric()}\cr
Optional grid to use for the finite differences. If \code{NULL} (the default), the argument grid of each functional
column is used. For \code{tfd_irreg} inputs, supplying \code{arg} interpolates the data to a common grid before
differentiating.
}
}

\examples{
task = tsk("fuel")
po_deriv = po("fda.derive", order = 1)
task_deriv = po_deriv$train(list(task))[[1L]]
task_deriv$data(cols = c("NIR", "UVVIS"))
}
\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}} -> \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{mlr3pipelines::PipeOpTaskPreprocSimple}} -> \code{PipeOpFDADerive}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PipeOpFDADerive-initialize}{\code{PipeOpFDADerive$new()}}
    \item \href{#method-PipeOpFDADerive-clone}{\code{PipeOpFDADerive$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpFDADerive-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpFDADerive-initialize}{}}}
\subsection{\code{PipeOpFDADerive$new()}}{
  Initializes a new instance of this Class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PipeOpFDADerive$new(id = "fda.derive", param_vals = list())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{(\code{character(1)})\cr
Identifier of resulting object, default \code{"fda.derive"}.}
      \item{\code{param_vals}}{(named \code{list()})\cr
List of hyperparameter settings, overwriting the hyperparameter settings that would
otherwise be set during construction. Default \code{list()}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpFDADerive-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpFDADerive-clone}{}}}
\subsection{\code{PipeOpFDADerive$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PipeOpFDADerive$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
