% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{phoneme}
\alias{phoneme}
\title{The phoneme dataset}
\format{
A data frame with 250 rows and 2 variables
}
\usage{
data(phoneme)
}
\description{
See \link{mlr_tasks_phoneme} for a description of the dataset.
}
\keyword{datasets}
\keyword{internal}
