% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional}
\alias{functional}
\title{Creates an object of class 'functional'}
\usage{
functional(args, values, ids)
}
\arguments{
\item{args}{(\code{numeric})}

\item{value}{(\code{numeric()})

export}
}
\description{
Creates an object of class functional.
}
