% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional}
\alias{functional}
\title{Creates a functional vector.}
\usage{
functional(args, values, ids)
}
\arguments{
\item{args}{(\code{numeric()}) Arguments of the functionals.}

\item{ids}{(\code{atomic} vector) Ids assigning the the (arg, value) combinations to a id.}

\item{value}{(\code{numeric()}) Values of the functionals.}
}
\description{
Creates an vector of class \code{functional} that contains function evaluations
from a usually unknown function f.
A functional vector is a list, where each element is again a named list
(list(arg = ..., value = ...)) with two equally long vectors \code{arg} and \code{value} such that
\code{value[i] = f(arg[i])}.
}
\examples{

}
