% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3fda-package}
\alias{mlr3fda}
\alias{mlr3fda-package}
\title{mlr3fda: What the Package Does (One Line, Title Case)}
\description{
Extends mlr3 with the functional data type.
Lang M, Binder M, Richter J, Schratz P, Pfisterer F, Coors S, Au Q, Casalicchio G, Kotthoff L, Bischl B (2019).
\dQuote{mlr3: A modern object-oriented machine learning framework in R.}
\emph{Journal of Open Source Software}.
\doi{10.21105/joss.01903}, \url{https://joss.theoj.org/papers/10.21105/joss.01903}.
}
\author{
\strong{Maintainer}: Sebastian Fischer \email{sebf.fischer@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
