% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpFDACatch22.R
\name{mlr_pipeops_fda.catch22}
\alias{mlr_pipeops_fda.catch22}
\alias{PipeOpFDACatch22}
\title{Catch22 Feature Extraction}
\description{
This \code{PipeOp} extracts the 22 (or 24) canonical time series characteristics (catch22) from functional columns.
For more details, see \code{\link[Rcatch22:catch22_all]{Rcatch22::catch22_all()}}, which is called internally on each curve.

The catch22 set is a low-redundancy subset of the \pkg{hctsa} features, selected for their performance across a
diverse collection of time series classification tasks, but applicable as general-purpose features for other tasks
such as regression.

For other time series feature extractors, see \code{\link{PipeOpFDATsfeatures}}.
}
\section{Parameters}{

The parameters are the parameters inherited from \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{PipeOpTaskPreprocSimple}},
as well as the following parameters:
\itemize{
\item \code{catch24} :: \code{logical(1)}\cr
If \code{TRUE}, additionally compute the mean and standard deviation (the catch24 set), yielding 24 features instead
of 22. Default is \code{FALSE}.
}
}

\section{Naming}{

The new names generally append a \verb{_\{feature\}} to the corresponding column name.
If a column was called \code{"x"} and the feature is \code{"DN_HistogramMode_5"}, the corresponding new column will
be called \code{"x_DN_HistogramMode_5"}.
}

\examples{
\dontshow{if (requireNamespace("Rcatch22", quietly = TRUE)) withAutoprint(\{ # examplesIf}
task = tsk("fuel")
po_catch22 = po("fda.catch22")
task_catch22 = po_catch22$train(list(task))[[1L]]
task_catch22$data()
\dontshow{\}) # examplesIf}
}
\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}} -> \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{mlr3pipelines::PipeOpTaskPreprocSimple}} -> \code{PipeOpFDACatch22}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PipeOpFDACatch22-initialize}{\code{PipeOpFDACatch22$new()}}
    \item \href{#method-PipeOpFDACatch22-clone}{\code{PipeOpFDACatch22$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpFDACatch22-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpFDACatch22-initialize}{}}}
\subsection{\code{PipeOpFDACatch22$new()}}{
  Initializes a new instance of this Class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PipeOpFDACatch22$new(id = "fda.catch22", param_vals = list())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{(\code{character(1)})\cr
Identifier of resulting object, default \code{"fda.catch22"}.}
      \item{\code{param_vals}}{(named \code{list()})\cr
List of hyperparameter settings, overwriting the hyperparameter settings that would
otherwise be set during construction. Default \code{list()}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpFDACatch22-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpFDACatch22-clone}{}}}
\subsection{\code{PipeOpFDACatch22$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PipeOpFDACatch22$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
