% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpFDAIntegrate.R
\name{mlr_pipeops_fda.integrate}
\alias{mlr_pipeops_fda.integrate}
\alias{PipeOpFDAIntegrate}
\title{Functional Integral Features}
\description{
Computes the definite integral of functional features via \code{\link[tf:tf_integrate]{tf::tf_integrate()}}.
The integral summarizes each curve into a single scalar, namely the (signed) area under the curve over its domain.

By default the integral is taken over the full domain of each curve. The \code{lower} and \code{upper} parameters restrict
the integration to a window. The same operation is applied during training and prediction.
}
\section{Parameters}{

The parameters are the parameters inherited from \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{PipeOpTaskPreprocSimple}},
as well as the following parameters:
\itemize{
\item \code{lower} :: \code{numeric(1)}\cr
The left boundary of the integration window. If not set, the domain start of each curve is used.
\item \code{upper} :: \code{numeric(1)}\cr
The right boundary of the integration window. If not set, the domain end of each curve is used.
}
}

\section{Naming}{

The new names generally append a \verb{_integral} to the corresponding column name.
If a column was called \code{"x"}, the corresponding new column will be called \code{"x_integral"}.
}

\examples{
task = tsk("fuel")
po_integrate = po("fda.integrate")
task_integrate = po_integrate$train(list(task))[[1L]]
task_integrate$data(cols = c("NIR_integral", "UVVIS_integral"))
}
\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}} -> \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{mlr3pipelines::PipeOpTaskPreprocSimple}} -> \code{PipeOpFDAIntegrate}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PipeOpFDAIntegrate-initialize}{\code{PipeOpFDAIntegrate$new()}}
    \item \href{#method-PipeOpFDAIntegrate-clone}{\code{PipeOpFDAIntegrate$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpFDAIntegrate-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpFDAIntegrate-initialize}{}}}
\subsection{\code{PipeOpFDAIntegrate$new()}}{
  Initializes a new instance of this Class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PipeOpFDAIntegrate$new(id = "fda.integrate", param_vals = list())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{(\code{character(1)})\cr
Identifier of resulting object, default \code{"fda.integrate"}.}
      \item{\code{param_vals}}{(named \code{list()})\cr
List of hyperparameter settings, overwriting the hyperparameter settings that would
otherwise be set during construction. Default \code{list()}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpFDAIntegrate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpFDAIntegrate-clone}{}}}
\subsection{\code{PipeOpFDAIntegrate$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PipeOpFDAIntegrate$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
