% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr_fuel.R
\name{mlr_tasks_fuel}
\alias{mlr_tasks_fuel}
\title{Fuel Regression Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskRegr}.
}
\description{
This dataset contains two functional covariates and one scalar covariate. The goal is
to predict the heat value of some fuel based on the ultraviolet radiation spectrum and
infrared ray radiation and one scalar column called h2o.

This is a subset of the full dataset, which is contained in the package \code{FDboost}.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_tasks$get("fuel")
tsk("fuel")
}\if{html}{\out{</div>}}
}

\references{
Brockhaus, Sarah, Scheipl, Fabian, Hothorn, Torsten, Greven, Sonja (2015).
\dQuote{The functional linear array model.}
\emph{Statistical Modelling}, \bold{15}(3), 279--300.
}
