% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskClassif_phoneme.R
\name{mlr_tasks_phoneme}
\alias{mlr_tasks_phoneme}
\title{Phoneme Classification Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskClassif}.
}
\description{
The task contains a single functional covariate and 5 equally big classes (aa, ao, dcl, iy, sh).
The aim is to predict the class of the phoneme in the functional, which is a
log-periodogram.\cr
This is a subset of the full dataset, which is contained in the package \code{fda.usc}.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_tasks$get("phoneme")
tsk("phoneme")
}\if{html}{\out{</div>}}
}

\references{
Ferraty, Frédric, Vieu, Philippe (2003).
\dQuote{Curves discrimination: a nonparametric functional approach.}
\emph{Computational Statistics & Data Analysis}, \bold{44}(1-2), 161--173.
}
